# src/openapi.yaml
openapi: "3.1.0"
info:
  title: WokAPI
  description: |
    WokAPI is the platform API for the WokSpec ecosystem — providing authentication,
    developer API tokens, billing, and AI routing.

    ## Authentication

    WokAPI supports two authentication methods:

    ### Session auth (cookie)
    Used by the WokSpec dashboard and web apps. Authenticate via GitHub or Google OAuth
    at `/v1/auth/github` or `/v1/auth/google`. The session cookie `wokspec_session` is
    set automatically.

    ### API key auth (Bearer token)
    For developer and programmatic access. Generate keys at `dashboard.wokspec.org/tokens`.

    ```
    Authorization: Bearer wok_live_a1b2c3d4...
    ```

    Keys are prefixed `wok_live_` (production) or `wok_test_` (sandbox).

    ## Rate limits

    All responses include rate limit headers:
    - `X-RateLimit-Limit` — requests allowed per minute
    - `X-RateLimit-Remaining` — requests remaining in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    | Plan | Requests/month | Requests/minute |
    |------|---------------|-----------------|
    | Free | 1,000 | 10 |
    | Pro | 100,000 | 100 |
    | Enterprise | Unlimited | 1,000 |

  version: "1.0.0"
  contact:
    name: WokSpec
    url: https://wokspec.org
    email: hello@wokspec.org
servers:
  - url: https://api.wokspec.org
    description: Production
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: wok_live_xxx
    SessionAuth:
      type: apiKey
      in: cookie
      name: wokspec_session
...
paths:
...
  /v1/billing/subscribe:
    post:
      summary: Create subscription checkout
      description: Creates a Stripe checkout session for upgrading to Pro or Enterprise.
      tags: [billing]
      security:
        - SessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                plan:
                  type: string
                  enum: [pro, enterprise]
                  default: pro
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      checkoutUrl:
                        type: string
                        format: uri
                      sessionId:
                        type: string
  /v1/billing/portal:
    post:
      summary: Get billing portal URL
      description: Returns a Stripe customer portal URL for managing subscriptions.
      tags: [billing]
      security:
        - SessionAuth: []
      responses:
        "200":
          description: Portal URL
...